🚚 Rename:  types/popper과 utils/popper 제거 및 Popper 폴더로 코드 이동, types/placement와 utils/placement 생성
