+++ src/components/data-display/Tour/Tour.stories.tsx
@@ -13,7 +13,7 @@ import { CloseIcon } from '@/components/icons/CloseIcon';
 import { ArrowLeftIcon } from '@/components/icons/ArrowLeftIcon';
 import { ArrowRightIcon } from '@/components/icons/ArrowRightIcon';
 import { StyleType } from '@/types/style';
-import { PlacementType } from '@/types/popper';
+import { PlacementType } from '@/types/placement';
 import { RadioGroup } from '@/components/data-entry/RadioGroup';
 import { Radio } from '@/components/data-entry/Radio';
 import { Label } from '@/components/data-entry/Label';
