+++ src/components/data-entry/Select/Select.constants.ts
@@ -1,10 +1,10 @@
-import { OriginType } from '@/types/popper';
+import { PopperProps } from '@/components/_share/Popper';
 
-export const ANCHOR_ORIGIN: OriginType = {
+export const ANCHOR_ORIGIN: PopperProps['anchorOrigin'] = {
   horizontal: 'center',
   vertical: 'bottom'
 };
-export const MENU_ORIGIN: OriginType = {
+export const MENU_ORIGIN: PopperProps['popperOrigin'] = {
   horizontal: 'center',
   vertical: 'top'
 };
