+++ src/types/placement.ts
@@ -1,0 +1,18 @@
+export type PlacementType =
+  | 'top-start'
+  | 'top'
+  | 'top-end'
+  | 'bottom-start'
+  | 'bottom'
+  | 'bottom-end'
+  | 'left-start'
+  | 'left'
+  | 'left-end'
+  | 'right-start'
+  | 'right'
+  | 'right-end';
+
+export type OriginType = {
+  horizontal: 'center' | 'left' | 'right' | number;
+  vertical: 'bottom' | 'center' | 'top' | number;
+};
